requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1

# LangChain/LangGraph + Google GenAI
langchain==0.2.16
langgraph==0.2.27
langchain-google-genai==2.0.4
google-generativeai==0.7.2
langchain-core==0.2.38

# Azure Search + Identity
azure-search-documents==11.6.0b4
azure-identity==1.17.1
azure-core==1.30.2

# Data & utils
pandas==2.2.2
openpyxl==3.1.5
httpx==0.27.0
